+++ s3transfer/manager.py
@@ -12,6 +12,7 @@
 # language governing permissions and limitations under the License.
 import copy
 import logging
+import re
 import threading
 
 from botocore.compat import six
@@ -70,7 +71,7 @@ class TransferConfig(object):
             processing a call to a TransferManager method. Processing a
             call usually entails determining which S3 API requests that need
             to be enqueued, but does **not** entail making any of the
-            S3 API data transfering requests needed to perform the transfer.
+            S3 API data transferring requests needed to perform the transfer.
             The threads controlled by ``max_request_concurrency`` is
             responsible for that.
 
@@ -98,7 +99,7 @@ class TransferConfig(object):
             will be tried upon errors with downloading an object in S3. Note
             that these retries account for errors that occur when streamming
             down the data from s3 (i.e. socket errors and read timeouts that
-            occur after recieving an OK response from s3).
+            occur after receiving an OK response from s3).
             Other retryable exceptions such as throttling errors and 5xx errors
             are already retried by botocore (this default is 5). The
             ``num_download_attempts`` does not take into account the
@@ -164,6 +165,7 @@ class TransferManager(object):
         'ContentEncoding',
         'ContentLanguage',
         'ContentType',
+        'ExpectedBucketOwner',
         'Expires',
         'GrantFullControl',
         'GrantRead',
@@ -177,6 +179,8 @@ class TransferManager(object):
         'SSECustomerKey',
         'SSECustomerKeyMD5',
         'SSEKMSKeyId',
+        'SSEKMSEncryptionContext',
+        'Tagging',
         'WebsiteRedirectLocation'
     ]
 
@@ -188,14 +192,25 @@ class TransferManager(object):
         'CopySourceSSECustomerAlgorithm',
         'CopySourceSSECustomerKey',
         'CopySourceSSECustomerKeyMD5',
-        'MetadataDirective'
+        'MetadataDirective',
+        'TaggingDirective',
     ]
 
     ALLOWED_DELETE_ARGS = [
         'MFA',
         'VersionId',
         'RequestPayer',
+        'ExpectedBucketOwner'
     ]
+
+    VALIDATE_SUPPORTED_BUCKET_VALUES = True
+
+    _UNSUPPORTED_BUCKET_PATTERNS = {
+        'S3 Object Lambda': re.compile(
+            r'^arn:(aws).*:s3-object-lambda:[a-z\-0-9]+:[0-9]{12}:'
+            r'accesspoint[/:][a-zA-Z0-9\-]{1,63}'
+        ),
+    }
 
     def __init__(self, client, config=None, osutil=None, executor_cls=None):
         """A transfer manager interface for Amazon S3
@@ -261,6 +276,14 @@ class TransferManager(object):
 
         self._register_handlers()
 
+    @property
+    def client(self):
+        return self._client
+
+    @property
+    def config(self):
+        return self._config
+
     def upload(self, fileobj, bucket, key, extra_args=None, subscribers=None):
         """Uploads a file to S3
 
@@ -292,6 +315,7 @@ class TransferManager(object):
         if subscribers is None:
             subscribers = []
         self._validate_all_known_args(extra_args, self.ALLOWED_UPLOAD_ARGS)
+        self._validate_if_bucket_supported(bucket)
         call_args = CallArgs(
             fileobj=fileobj, bucket=bucket, key=key, extra_args=extra_args,
             subscribers=subscribers
@@ -334,6 +358,7 @@ class TransferManager(object):
         if subscribers is None:
             subscribers = []
         self._validate_all_known_args(extra_args, self.ALLOWED_DOWNLOAD_ARGS)
+        self._validate_if_bucket_supported(bucket)
         call_args = CallArgs(
             bucket=bucket, key=key, fileobj=fileobj, extra_args=extra_args,
             subscribers=subscribers
@@ -387,6 +412,9 @@ class TransferManager(object):
         if source_client is None:
             source_client = self._client
         self._validate_all_known_args(extra_args, self.ALLOWED_COPY_ARGS)
+        if isinstance(copy_source, dict):
+            self._validate_if_bucket_supported(copy_source.get('Bucket'))
+        self._validate_if_bucket_supported(bucket)
         call_args = CallArgs(
             copy_source=copy_source, bucket=bucket, key=key,
             extra_args=extra_args, subscribers=subscribers,
@@ -421,11 +449,25 @@ class TransferManager(object):
         if subscribers is None:
             subscribers = []
         self._validate_all_known_args(extra_args, self.ALLOWED_DELETE_ARGS)
+        self._validate_if_bucket_supported(bucket)
         call_args = CallArgs(
             bucket=bucket, key=key, extra_args=extra_args,
             subscribers=subscribers
         )
         return self._submit_transfer(call_args, DeleteSubmissionTask)
+
+    def _validate_if_bucket_supported(self, bucket):
+        # s3 high level operations don't support some resources
+        # (eg. S3 Object Lambda) only direct API calls are available
+        # for such resources
+        if self.VALIDATE_SUPPORTED_BUCKET_VALUES:
+            for resource, pattern in self._UNSUPPORTED_BUCKET_PATTERNS.items():
+                match = pattern.match(bucket)
+                if match:
+                    raise ValueError(
+                        'TransferManager methods do not support %s '
+                        'resource. Use direct client calls instead.' % resource
+                    )
 
     def _validate_all_known_args(self, actual, allowed):
         for kwarg in actual:
@@ -600,7 +642,7 @@ class TransferCoordinatorController(object):
             self._tracked_transfer_coordinators.add(transfer_coordinator)
 
     def remove_transfer_coordinator(self, transfer_coordinator):
-        """Remove a transfer coordinator from cancelation consideration
+        """Remove a transfer coordinator from cancellation consideration
 
         Typically, this method is invoked by the transfer coordinator itself
         to remove its self when it completes its transfer.