+++ s3transfer/upload.py
@@ -85,8 +85,8 @@ class InterruptReader(object):
             raise self._transfer_coordinator.exception
         return self._fileobj.read(amount)
 
-    def seek(self, where):
-        self._fileobj.seek(where)
+    def seek(self, where, whence=0):
+        self._fileobj.seek(where, whence)
 
     def tell(self):
         return self._fileobj.tell()
@@ -490,10 +490,12 @@ class UploadSubmissionTask(SubmissionTask):
         'SSECustomerAlgorithm',
         'SSECustomerKeyMD5',
         'RequestPayer',
+        'ExpectedBucketOwner'
     ]
 
     COMPLETE_MULTIPART_ARGS = [
-        'RequestPayer'
+        'RequestPayer',
+        'ExpectedBucketOwner'
     ]
 
     def _get_upload_input_manager_cls(self, transfer_future):