+++ s3transfer/utils.py
@@ -239,6 +239,8 @@ class CountCallbackInvoker(object):
 
 
 class OSUtils(object):
+    _MAX_FILENAME_LEN = 255
+
     def get_file_size(self, filename):
         return os.path.getsize(filename)
 
@@ -300,7 +302,11 @@ class OSUtils(object):
         return False
 
     def get_temp_filename(self, filename):
-        return filename + os.extsep + random_file_extension()
+        suffix = os.extsep + random_file_extension()
+        path = os.path.dirname(filename)
+        name = os.path.basename(filename)
+        temp_filename = name[:self._MAX_FILENAME_LEN - len(suffix)] + suffix 
+        return os.path.join(path, temp_filename)
 
     def allocate(self, filename, size):
         try:
@@ -357,9 +363,9 @@ class DeferredOpenFile(object):
         self._open_if_needed()
         self._fileobj.write(data)
 
-    def seek(self, where):
+    def seek(self, where, whence=0):
         self._open_if_needed()
-        self._fileobj.seek(where)
+        self._fileobj.seek(where, whence)
 
     def tell(self):
         if self._fileobj is None:
@@ -419,6 +425,8 @@ class ReadFileChunk(object):
         self._size = self._calculate_file_size(
             self._fileobj, requested_size=chunk_size,
             start_byte=self._start_byte, actual_file_size=full_file_size)
+        # _amount_read represents the position in the chunk and may exceed
+        # the chunk size, but won't allow reads out of bounds.
         self._amount_read = 0
         self._callbacks = callbacks
         if callbacks is None:
@@ -467,10 +475,11 @@ class ReadFileChunk(object):
         return min(max_chunk_size, requested_size)
 
     def read(self, amount=None):
+        amount_left = max(self._size - self._amount_read, 0)
         if amount is None:
-            amount_to_read = self._size - self._amount_read
+            amount_to_read = amount_left
         else:
-            amount_to_read = min(self._size - self._amount_read, amount)
+            amount_to_read = min(amount_left, amount)
         data = self._fileobj.read(amount_to_read)
         self._amount_read += len(data)
         if self._callbacks is not None and self._callbacks_enabled:
@@ -493,13 +502,29 @@ class ReadFileChunk(object):
     def disable_callback(self):
         self._callbacks_enabled = False
 
-    def seek(self, where):
-        self._fileobj.seek(self._start_byte + where)
+    def seek(self, where, whence=0):
+        if whence not in (0, 1, 2):
+            # Mimic io's error for invalid whence values
+            raise ValueError(
+                "invalid whence (%s, should be 0, 1 or 2)" % whence)
+
+        # Recalculate where based on chunk attributes so seek from file
+        # start (whence=0) is always used
+        where += self._start_byte
+        if whence == 1:
+            where += self._amount_read
+        elif whence == 2:
+            where += self._size
+
+        self._fileobj.seek(max(where, self._start_byte))
         if self._callbacks is not None and self._callbacks_enabled:
             # To also rewind the callback() for an accurate progress report
+            bounded_where = max(min(where - self._start_byte, self._size), 0)
+            bounded_amount_read = min(self._amount_read, self._size)
+            amount = bounded_where - bounded_amount_read
             invoke_progress_callbacks(
-                self._callbacks, bytes_transferred=where - self._amount_read)
-        self._amount_read = where
+                self._callbacks, bytes_transferred=amount)
+        self._amount_read = max(where - self._start_byte, 0)
 
     def close(self):
         if self._close_callbacks is not None and self._callbacks_enabled:
@@ -567,7 +592,7 @@ class TaskSemaphore(object):
             needed for API compatibility with the SlidingWindowSemaphore
             implementation.
         :param block: If True, block until it can be acquired. If False,
-            do not block and raise an exception if cannot be aquired.
+            do not block and raise an exception if cannot be acquired.
 
         :returns: A token (can be None) to use when releasing the semaphore
         """