+++ tests/unit/test_download.py
@@ -58,7 +58,7 @@ class WriteCollector(object):
         self._pos = 0
         self.writes = []
 
-    def seek(self, pos):
+    def seek(self, pos, whence=0):
         self._pos = pos
 
     def write(self, data):
@@ -684,7 +684,7 @@ class TestGetObjectTask(BaseTaskTest):
 
         self.stubber.assert_no_pending_responses()
         expected_contents = []
-        # This is the content intially read in before the retry hit on the
+        # This is the content initially read in before the retry hit on the
         # second read()
         expected_contents.append((0, bytes(self.content[0:1])))